feat: all cubes are rendered using instancing
Still white, but they're all rendered. Commit is, once again, huge
because of changes that are connected.

Data resides in one of two "backing buffers" depending on their writing
pattern. All instance data ends up in a third "round robin" buffer that
is divided into as many parts as there are backbuffers. Data is copied
from the backing buffer to the round robin buffer before rendering.