feat: bind specular texture
It can't be in the same descriptor set as the diffuse texture since
there is no hint to which texture is the diffuse and which texture is
the normal, so they have to be bound separately.

If the descriptor sets were updated in RendererVulkan::Render they could
be in the same set, but that would require too big of a rewrite.